* main: (29 commits)
  add layer style: outer glow (#890)
  clear jni exception (#891)
  Fix memory leak. (#893)
  Add a RemoveAll() method to PAGDiskCache. (#892)
  Update the implementation of PAGImageView. (#887)
  add LayerStyle: stroke(codec) (#885)
  Fix include. (#888)
  Disable mipmap temporary. (#886)
  Fix: update ImageBitmap compatibility check for Safari 15- (#884)
  Add support for saving static time ranges info into the SequenceFile. (#883)
  Unified the setting of valueAnimator‘s progress. (#880)
  reset repeat count to zero when call setProgress (#881)
  Fixing the WebMask texture type upload error that causes WeChat crash. (#877)
  Move the PAGDeocder implementation to the C++ side. (#876)
  Fix: Add fallback for OffscreenCanvas in built-in iOS 16.4 AppleWebKit browsers (#875)
  Add support for decoding images into HardwareBuffers directly on the android platform. (#873)
  Refactor PAGView on WeChat miniprogram (#871)
  Fixe problem of PAGDecoder, make the cache storage of PAGImagView synchronized to ensure data security. (#870)
  update android tools version (#872)
  Add support for creating offscreen PAGSurface on web and qt platforms. (#869)
  ...

# Conflicts:
#	src/platform/ios/PAGImageView.mm

+++ src/codec/utils/EncodeStream.cpp
@@ -149,7 +149,7 @@ void EncodeStream::writeByteData(const pag::ByteData* byteData) {
 }
 
 void EncodeStream::writeUTF8String(const std::string& text) {
-  auto textLength = text.size();
+  auto textLength = static_cast<uint32_t>(text.size());
   if (checkCapacity(textLength + 1)) {
     memcpy(bytes + _position, text.c_str(), textLength + 1);
     positionChanged(static_cast<off_t>(textLength) + 1);
